% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{finish_position}
\alias{finish_position}
\title{Retrieve competitor finish position}
\usage{
finish_position(eventId, compEventId)
}
\arguments{
\item{eventId}{Database ID of the event}

\item{compEventId}{Event Competitor ID}
}
\description{
Retrieve competitor finish position
}
\examples{
event_finish_positions( 1175769, 10962625 )
}
\keyword{competitors}

