% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fitLogistic}
\alias{fitLogistic}
\title{Fit Logistic Curve}
\usage{
fitLogistic(data, y = "N", time = "Day", N0 = 1, K = 200, r0 = 0.75)
}
\description{
Fit Logistic Curve
}
\examples{
data(single)

plot(aurelia ~ Day, data = single, col = 1, type = "o")
lines(caudatum ~ Day, data = single, col = 2, type = "o")
legend("bottomright", legend = c("aurelia", "caudatum"), col = 1:2, pch = 1, lty = 1)

fit1 <- fitLogistic(single, y = "aurelia", time = "Day", 1, 200, .75)
fit2 <- fitLogistic(single, y = "caudatum", time = "Day", 1, 200, .75)

linesLogistic(fit1, lwd = 3)
linesLogistic(fit2, col = 2, lwd = 3)
}
