% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logistic}
\alias{logistic}
\title{Logistic function}
\usage{
logistic(x, N0, K, r0)
}
\arguments{
\item{x}{time}

\item{N0}{initial time}

\item{K}{carrying capacity}

\item{r0}{intrinsic growth rate}
}
\description{
This is the logistic function
which grows to a carrying capacity
}
\examples{
curve(logistic(x, .01, 1, 10))

}
