% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{scratched}
\alias{scratched}
\title{Retrieve competitor scratched property}
\usage{
scratched(eventId, compEventId)
}
\arguments{
\item{eventId}{Database ID of the event}

\item{compEventId}{Event Competitor ID}
}
\description{
Retrieve competitor scratched property
}
\examples{
scratched( 1175769, 10962625 )
}
\keyword{competitors}

