% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{scratchings}
\alias{scratchings}
\title{Retrieve competitor scratched property}
\usage{
scratchings(eventId)
}
\arguments{
\item{eventId}{Database ID of the event}
}
\description{
Retrieve competitor scratched property
}
\examples{
scratchings( 1175769 )
}
\keyword{competitors}

