% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{single}
\alias{single}
\title{Single separate paramecium growth}
\format{
Three columns:
\describe{
\item{Day}{Day of experiment}
\item{caudatum}{volume of P caudatum}
\item{aurelia}{volume of P aurelia}
}
}
\source{

}
\usage{
data(single)
}
\description{
Growth of two species of paramecium, P aurelia
and P caudatum, in separate petri dishes.
}
\examples{
data(single)
plot(aurelia ~ Day, data = single)
}
\keyword{data}
